refactor: 받아오는 id값 Article id에 맞춰 int로 형 변환

+++ src/main/java/techcourse/myblog/domain/ArticleRepository.java
@@ -24,9 +24,9 @@ public class ArticleRepository {
         articles.add(article);
     }
 
-    public Article find(long articleId) {
+    public Article find(int articleId) {
         return articles.stream()
-                .filter(article -> article.getId() == (int) articleId)
+                .filter(article -> article.getId() == articleId)
                 .findFirst()
                 .orElseThrow(IllegalArgumentException::new)
                 ;
@@ -37,9 +37,9 @@ public class ArticleRepository {
         articles.add(editedArticle);
     }
 
-    public void delete(long articleId) {
+    public void delete(int articleId) {
         Article articleToDelete = articles.stream()
-                .filter(article -> article.getId() == (int) articleId)
+                .filter(article -> article.getId() == articleId)
                 .findFirst()
                 .orElseThrow(IllegalArgumentException::new);
         articles.remove(articleToDelete);
